Fix `splitMessage`, remove `decryptMIMEMessage` (unused), rewrite message/utils in TS
