+++ lib/message/processMIME.js
@@ -1,4 +1,5 @@
-import { getSignature, verifyMessage, removeTrailingSpaces } from './utils';
+import { getSignature, removeTrailingSpaces } from './utils';
+import { verifyMessage } from './verify';
 import { VERIFICATION_STATUS, MAX_ENC_HEADER_LENGTH } from '../constants';
 import { serverTime } from '../serverTime';
 import { parseMail } from './parseMail';
