+++ lib/message/verify.js
@@ -1,112 +1,9 @@
-/* eslint-disable no-prototype-builtins */
-import { readToEnd } from '@openpgp/web-stream-tools';
-import {
-    createMessage,
-    PacketList,
-    verify,
-    sign,
-    enums,
-    readCleartextMessage,
-    Message,
-    CleartextMessage,
-    readMessage,
-    Signature,
-    readSignature,
-    unarmor
-} from '../openpgp';
+import { createMessage, verify, CleartextMessage } from '../openpgp';
 import { VERIFICATION_STATUS } from '../constants';
 import { serverTime } from '../serverTime';
+import { removeTrailingSpaces } from './utils';
 
 const { NOT_SIGNED, SIGNED_AND_VALID, SIGNED_AND_INVALID } = VERIFICATION_STATUS;
-
-/**
- * Remove trailing spaces and tabs from each line (separated by \n characters)
- */
-export const removeTrailingSpaces = (text) => {
-    return text
-        .split('\n')
-        .map((line) => {
-            let i = line.length - 1;
-            for (; i >= 0 && (line[i] === ' ' || line[i] === '\t'); i--);
-            return line.substr(0, i + 1);
-        })
-        .join('\n');
-};
-
-/**
- * Prepare message
- * @param {Object|Uint8Array|String} message
- * @return {Promise<Object>}
- */
-export async function getMessage(message) {
-    if (Message.prototype.isPrototypeOf(message)) {
-        return message;
-    }
-    if (Uint8Array.prototype.isPrototypeOf(message)) {
-        return readMessage({ binaryMessage: message });
-    }
-    return readMessage({ armoredMessage: message.trim() });
-}
-
-/**
- * Prepare signature
- * @param {String|Uint8Array|openpgp.signature.Signature} signature
- * @return {Promise<openpgp.signature.Signature>}
- */
-export async function getSignature(signature) {
-    if (Signature.prototype.isPrototypeOf(signature)) {
-        return signature;
-    }
-    if (Uint8Array.prototype.isPrototypeOf(signature)) {
-        return readSignature({ binarySignature: signature });
-    }
-    return readSignature({ armoredSignature: signature.trim() });
-}
-
-/**
- * Read a cleartext message from an armored message.
- * @param {String|openpgp.cleartext.CleartextMessage} message
- * @return {Promise<openpgp.cleartext.CleartextMessage>}
- */
-export async function getCleartextMessage(message) {
-    if (CleartextMessage.prototype.isPrototypeOf(message)) {
-        return message;
-    }
-    return readCleartextMessage({ cleartextMessage: message.trim() });
-}
-
-/**
- * Get a signed message from the given data.
- * Either `textData` or `binaryData` must be specified.
- * @param {Object} options - input for openpgp.sign
- * @param {String|ReadableStream<String>} textData - text data to sign
- * @param {Uint8Array|ReadableStream<Uint8Array>} binaryData - binary data to sign
- * @param {Boolean} stripTrailingSpaces - whether trailing spaces should be removed from `textData`
- * @returns Promise<{Message|Signature|MaybeStream<String>|MaybeStream<Uint8Array>}> signed message object, signature object, or corresponding serialised data
- * @throws on signing error
- */
-export async function signMessage({
-    textData,
-    binaryData,
-    stripTrailingSpaces,
-    date = serverTime(),
-    format = 'armored',
-    ...options
-}) {
-    const dataType = binaryData ? 'binary' : 'text';
-    const data = binaryData || (stripTrailingSpaces ? removeTrailingSpaces(textData) : textData); // throw if streamed text and stripTrailingSpaces enabled
-    const sanitizedOptions = {
-        ...options,
-        date,
-        format,
-        message: await createMessage({ [dataType]: data, date })
-    };
-
-    return sign(sanitizedOptions).catch((err) => {
-        console.error(err);
-        return Promise.reject(err);
-    });
-}
 
 /**
  * Extract information from the result of openpgp.verify
@@ -224,9 +121,9 @@ export async function verifyMessage({ textData, binaryData, stripTrailingSpaces,
  *     errors: Error[]|undefined - verification errors if all signatures are invalid
  * }
  */
- export async function verifyCleartextMessage({ cleartextMessage, date = serverTime(), ...options }) {
+export async function verifyCleartextMessage({ cleartextMessage, date = serverTime(), ...options }) {
     if (!(cleartextMessage instanceof CleartextMessage)) {
-        throw new Error('CleartextMessage expected.')
+        throw new Error('CleartextMessage expected.');
     }
     const sanitizedOptions = {
         ...options,
@@ -248,78 +145,3 @@ export async function verifyMessage({ textData, binaryData, stripTrailingSpaces,
             return Promise.reject(err);
         });
 }
-
-export async function splitMessage(message) {
-    const msg = await getMessage(message);
-
-    const keyFilter = (packet) => {
-        return (
-            packet.constructor.tag !== enums.packet.publicKeyEncryptedSessionKey &&
-            packet.constructor.tag !== enums.packet.signature &&
-            packet.constructor.tag !== enums.packet.symEncryptedSessionKey &&
-            packet.constructor.tag !== enums.packet.compressed &&
-            packet.constructor.tag !== enums.packet.literal &&
-            packet.constructor.tag !== enums.packet.symmetricallyEncrypted &&
-            packet.constructor.tag !== enums.packet.symEncryptedIntegrityProtected &&
-            packet.constructor.tag !== enums.packet.symEncryptedAEADProtected
-        );
-    };
-
-    const splitPackets = (packetList) => {
-        return Promise.all(
-            packetList.map((pack) => {
-                const newList = new PacketList();
-                newList.push(pack);
-                const data = newList.write(); // Uint8Array / String (ReadableStream)
-
-                // readToEnd is async and accepts Uint8Array/String
-                return readToEnd(data);
-            })
-        );
-    };
-
-    const asymmetric = await splitPackets(msg.packets.filterByTag(enums.packet.publicKeyEncryptedSessionKey));
-    const signature = await splitPackets(msg.packets.filterByTag(enums.packet.signature));
-    const symmetric = await splitPackets(msg.packets.filterByTag(enums.packet.symEncryptedSessionKey));
-    const compressed = await splitPackets(msg.packets.filterByTag(enums.packet.compressed));
-    const literal = await splitPackets(msg.packets.filterByTag(enums.packet.literal));
-    const encrypted = await splitPackets(
-        msg.packets.filterByTag(
-            enums.packet.symmetricallyEncrypted,
-            enums.packet.symEncryptedIntegrityProtected,
-            enums.packet.symEncryptedAEADProtected
-        )
-    );
-    const other = await splitPackets(msg.packets.filter(keyFilter));
-
-    return {
-        asymmetric,
-        signature,
-        symmetric,
-        compressed,
-        literal,
-        encrypted,
-        other
-    };
-}
-
-/**
- *  Prepare message body
- * @param {String} value
- * @return {Promise<String>}
- */
-export async function armorBytes(value) {
-    const bodyMessage = await getMessage(value);
-    return readToEnd(bodyMessage.armor());
-}
-
-/**
- * Dearmor a pgp encoded message.
- * @param {String} input
- * @return {Promise<Uint8Array>}
- */
-export const stripArmor = async (input) => {
-    const { data } = await unarmor(input);
-    const bytes = await readToEnd(data);
-    return bytes;
-};
