+++ lib/pmcrypto.js
@@ -27,16 +27,16 @@ export {
 
 export { decryptPrivateKey, decryptSessionKey } from './key/decrypt';
 export { encryptPrivateKey, encryptSessionKey } from './key/encrypt';
-export { decryptMessage, decryptMessageLegacy, decryptMIMEMessage } from './message/decrypt';
+export { default as decryptMessage } from './message/decrypt';
+export { default as decryptMessageLegacy } from './message/decryptLegacy';
 export { default as encryptMessage } from './message/encrypt';
+export { default as signMessage } from './message/sign';
+export { verifyMessage, verifyCleartextMessage } from './message/verify';
 
 export {
     getMessage,
     getSignature,
-    signMessage,
     splitMessage,
-    verifyMessage,
-    verifyCleartextMessage,
     getCleartextMessage,
     armorBytes,
     stripArmor
